% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_estimation.R
\name{benchmark_performance}
\alias{benchmark_performance}
\title{Quantifies user thread speed}
\usage{
benchmark_performance()
}
\value{
Proportion difference between the mean observed execution time and the mean estimated execution time.
}
\description{
This function benchmarks the performance of a network analysis function
by comparing observed execution times with estimated times derived from a predictive model.
It generates networks with specified size, sparsity, and repeats the process to average results.
}
\keyword{internal}
