% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betweeness_centrality.R
\name{betweenness_wei}
\alias{betweenness_wei}
\title{Calculate Betweenness Centrality for Weighted Networks}
\usage{
betweenness_wei(W, validate = TRUE)
}
\arguments{
\item{W}{A square numeric matrix representing the weighted network, where higher weights imply stronger connections.}

\item{validate}{Whether to validate the input matrix.}
}
\value{
A numeric vector containing the betweenness centrality for each node in the network.
}
\description{
This function computes the betweenness centrality for each node in a weighted network represented by a matrix.
Betweenness centrality measures the extent to which a node lies on paths between other nodes.
Nodes with higher betweenness centrality play a more significant role in the communication/interaction within the network.
The function assumes that a zero in the matrix represents no connection and infinities represent disconnected paths.
}
\examples{
W<- matrix(c(0, 2, 1, 0, 2, 0, 3, 5, 1, 3, 0, 6, 0, 5, 6, 0), nrow = 4, byrow = TRUE)
betweenness_wei(W)
}
