% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_rich_club_nodes.R
\name{classify_rich_club_nodes}
\alias{classify_rich_club_nodes}
\title{Classify Rich Club Nodes}
\usage{
classify_rich_club_nodes(data_frame)
}
\arguments{
\item{data_frame}{A data frame containing the network data with columns:
"subject", "row", "col", "edge_strength", "self_con".}
}
\value{
A vector of the top 10 rich club nodes identified within the network.
}
\description{
This function identifies the top 10 rich club nodes in a network
based on their edge strengths, using edges produced by the function \code{process_matrices()}
and classified using the function \code{classify_connections()}. It specifically
focusing on edges between distinct nodes marked as "Network-Connected".
}
\examples{
data_frame <- data.frame(subject = c(1, 1, 2, 2),
                         row = c("A", "B", "A", "C"),
                         col = c("B", "A", "C", "A"),
                         edge_strength = c(2, 2, 3, 3),
                         self_connectivity = c("Network-Connected",
                         "Network-Connected", "Network-Connected",
                         "Network-Connected"))
classify_rich_club_nodes(data_frame)

}
