% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich_club.R
\name{compare_rich_club}
\alias{compare_rich_club}
\title{Rich Club Coefficient Comparison}
\usage{
compare_rich_club(
  mat_array,
  subjects,
  grouping_list,
  plot = FALSE,
  permutation = FALSE,
  n_rand = 100,
  p_adj = TRUE
)
}
\arguments{
\item{mat_array}{An array of connectivity matrices where each slice corresponds to a different subject.}

\item{subjects}{A vector of subject identifiers.}

\item{grouping_list}{A list specifying how subjects should be grouped for analysis.}

\item{plot}{A boolean indicating whether to generate plots of the results.}

\item{permutation}{A boolean to choose whether permutation tests should be used instead of ANOVA for statistical comparison.}

\item{n_rand}{The number of random permutations to use if permutation is TRUE.}

\item{p_adj}{A boolean indicating whether significant differences between groups on the graph should be controlled for multiple comparisons.}
}
\value{
A list containing statistical analysis results and optionally a plot, depending on the input parameters.
}
\description{
Computes and compares rich club coefficients for multiple subjects grouped by given criteria.
}
\examples{
data_dir <- system.file("extdata", package = "NetAnalyseR")
subjects <- c("A", "B", "C", "D")
grouping <- list(AB = c("A" , "B"), CD = c("C", "D"))
file_convention <- ".csv"
output <- process_matrices(data_dir, subjects, file_convention)
compare_rich_club(mat_array = output$matrices,
                  subjects = output$subjects,
                  n_rand = 100,
                  grouping_list = grouping,
                  plot = T)



}
