% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_global_metrics.R
\name{compute_global_metrics}
\alias{compute_global_metrics}
\title{Compute Global Metrics for Brain Network Data}
\usage{
compute_global_metrics(matrices_array, global_metrics, subject_names = NULL)
}
\arguments{
\item{matrices_array}{An array where each slice represents a connectivity matrix for a subject.}

\item{global_metrics}{A vector of strings specifying the global metrics to be calculated.}

\item{subject_names}{An optional vector of subject identifiers. If not provided, subjects will be named
sequentially as Subject1, Subject2, etc.}
}
\value{
A data frame containing the calculated global metrics for each subject, with each row corresponding
to a subject and columns for the subject identifier and the calculated metric values.
}
\description{
This function calculates specified global metrics for a given set of brain network matrices
for multiple subjects. It validates the requested global metrics against a list of supported metrics, warns
about invalid metrics, and computes the specified metrics for each subject's brain network.
}
\examples{
W <- matrix(c(0, 2, 1, 0, 2, 0, 3, 5, 1, 3, 0, 6, 0, 5, 6, 0), nrow = 4, byrow = TRUE)
valid_global_metrics <- c("characteristic_path_length",
                          "global_clustering_coefficient_wei",
                          "global_efficiency_wei",
                          "inter_node",
                          "intra_node",
                          "network_density",
                          "normalised_clustering_coefficient",
                          "normalised_characteristic_path_length",
                          "small_worldness")
subject_names <- c("Subject1")
global_df <- compute_global_metrics(W, valid_global_metrics, subject_names)
}
