% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_structural_deviation.R
\name{compute_network_deviation}
\alias{compute_network_deviation}
\title{Compute Network Deviations}
\usage{
compute_network_deviation(
  data_frame,
  matrices_array,
  control_group,
  threshold = 3
)
}
\arguments{
\item{data_frame}{A data frame containing at least two columns: \code{group} and \code{subject} for participant identifiers. The}

\item{matrices_array}{A 3D array where each slice corresponds to a matrix for a specific participant.}

\item{control_group}{A character specifying the level of \code{group} which should be considered the control group.}

\item{threshold}{A numeric value defining the Z-score threshold for considering deviations as significant.}
}
\value{
A data frame that joins the input data_frame with computed network deviations for each case.
}
\description{
This function calculates network deviations for subjects based on the provided control and subject groups within a data frame.
It extracts control and subject matrices from a 3D array based on group classification in the data frame and then calculates network deviations
using a specified threshold.
}
\details{
Compute network deviations for a given dataset and matrices array
}
\examples{
# data_frame is a data frame with columns 'group' and 'subject'.
# matrices_array is a 3D array where the third dimension corresponds to individuals in the data_frame.
# Assuming network_deviation is a predefined function that takes a threshold as an argument.
result <- compute_network_deviation(data_frame, matrices_array, 2.5)
}
\references{
Gugger, J. J., Sinha, N., Huang, Y., Walter, A. E., Lynch, C., Kalyani, P., Smyk, N., Sandsmark, D., Diaz-Arrastia, R., & Davis, K. A. (2023). Structural brain network deviations predict recovery after traumatic brain injury. NeuroImage clinical, 38, 103392-103392. https://doi.org/10.1016/j.nicl.2023.103392
}
