% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_structural_deviation.R
\name{compute_nodal_network_deviation}
\alias{compute_nodal_network_deviation}
\title{Compute Nodal Network Deviations}
\usage{
compute_nodal_network_deviation(
  data_frame,
  matrices_array,
  control_group,
  threshold = 3
)
}
\arguments{
\item{data_frame}{A data frame where each row corresponds to a single subject. Typically produced using \code{compute_nodal_metrics()}. Must have a column \code{group} and column \code{subject}.}

\item{matrices_array}{A 3D array where each slice corresponds to a connectivity matrix for a subject.}

\item{control_group}{A character specifying the level of \code{group} which should be considered the control group.}

\item{threshold}{A numeric value defining the Z-score threshold for considering deviations significant.}
}
\value{
A modified version of the original data frame that includes nodal deviation calculations for each subject.
}
\description{
This function computes nodal network deviations for all nodes for all subjects in a dataset.
The nodal deviation is calculated as the number of edges of a given node which have a Z score which significantly differ from the
function to calculate deviations and merges these results back into the original data frame.
}
\details{
Compute nodal network deviations and merge with original data frame
}
\examples{
data_dir <- system.file("extdata", package = "NetAnalyseR")
subjects <- c("A", "B", "C", "D")
file_convention <- ".csv"
output <- process_matrices(data_dir, subjects, file_convention)
nodal_metrics <- c("local_efficiency_wei")
nodal_df <- compute_nodal_metrics(output$matrices, nodal_metrics, output$subjects)
grouping_list <- list(
  control = c("A", "B"),
  treatment = c("C", "D")
)
allocated_df <- allocate_groups(nodal_df, grouping_list)
nodal_deviation <- compute_nodal_network_deviation(allocated_df, output$matrices, "control")
}
