% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{floydWarshallRcpp}
\alias{floydWarshallRcpp}
\title{Floyd-Warshall Algorithm in R}
\arguments{
\item{inputMatrix}{A NumericMatrix representing the adjacency matrix of the graph where the element at the ith row
and jth column represents the weight of the edge from vertex i to vertex j. If there is no edge between vertex i and vertex j,
the value should be 0 for i != j.}
}
\value{
A NumericMatrix where the element at the ith row and jth column represents the shortest distance from vertex i
to vertex j in the input graph.
}
\description{
This function implements the Floyd-Warshall algorithm using Rcpp for finding the shortest distances
between every pair of vertices in a given edge weighted directed Graph. It operates on a matrix representation of the graph.
}
\examples{
# Define the adjacency matrix for a graph
# Calculate shortest paths
shortestPaths <- floydWarshallRcpp(W)
print(shortestPaths)
}
