% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{generate_warning}
\alias{generate_warning}
\title{Generate a Warning Once}
\usage{
generate_warning(warning_var_name, warning_message)
}
\arguments{
\item{warning_var_name}{A string representing the name of the variable used to track whether the warning has been issued.}

\item{warning_message}{A string containing the warning message to be displayed.}
}
\value{
No return value; the function solely issues a warning and updates the global environment.
}
\description{
This function generates a specified warning message only once during the R session.
}
\examples{
generate_warning_once("warned_about_issue", "Attention: Specific issue has occurred!")
}
\keyword{internal}
