% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateRewiredMatrices}
\alias{generateRewiredMatrices}
\title{Generate a Series of Rewired Matrices}
\usage{
generateRewiredMatrices(initialMatrix, n)
}
\arguments{
\item{initialMatrix}{A numeric matrix representing the initial network's adjacency matrix.}

\item{n}{The number of rewired matrices to generate.}
}
\value{
A list of 'n' rewired matrices.
}
\description{
Starts with an initial network matrix, performs an initial 100 edge rewirings,
and then continues to rewire the network with 1 iteration for each subsequent step,
saving the state of the matrix after each single rewiring. This process is repeated
until 'n' rewired matrices are generated.
}
