% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{K_dis_clust}
\alias{K_dis_clust}
\title{K-Means Distance-Based Clustering}
\usage{
K_dis_clust(mat, k = floor(dim(mat)[1]/2))
}
\arguments{
\item{mat}{An adjacency matrix from which the shortest distance matrix will be computed.}

\item{k}{An integer specifying the number of clusters. Default is set to one-eighth of the number of rows in the matrix.}
}
\value{
A vector of cluster labels for each row in the input matrix.
}
\description{
Perform K-means clustering based on the shortest distance matrix derived from the input matrix.
}
\examples{
# Create a sample matrix
W <- matrix(c(0, 2, 1, 0, 2, 0, 3, 5, 1, 3, 0, 6, 0, 5, 6, 0), nrow = 4, byrow = TRUE)
# Perform K-means clustering
cluster_labels <- K_dis_clust(W)
print(cluster_labels)
}
