% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{label_prop_clust}
\alias{label_prop_clust}
\title{Weighted Semi-Synchronous Label Propagation}
\usage{
label_prop_clust(weighted_adj_matrix)
}
\arguments{
\item{weighted_adj_matrix}{A square matrix representing the weighted adjacency matrix of the graph.}
}
\value{
A vector of labels indicating the community assignment for each node.
}
\description{
Perform weighted semi-synchronous label propagation on a given
weighted adjacency matrix. Automatically determines the number of clusters to
generate.
}
\examples{
# Create a sample weighted adjacency matrix
W <- matrix(c(0, 2, 1, 0, 2, 0, 3, 5, 1, 3, 0, 6, 0, 5, 6, 0), nrow = 4, byrow = TRUE)
# Perform weighted semi-synchronous label propagation
labels <- label_prop_clust(W)
print(labels)
}
\references{
Cordasco, G, and L Gargano. “Community Detection via Semi-Synchronous Label Propagation Algorithms.” 2010 IEEE International Workshop on: Business Applications of Social Network Analysis (BASNA). IEEE, 2010. 1–8. Web.
}
