% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mat_gen}
\alias{mat_gen}
\title{Generate Symmetric Matrix with Random Values}
\usage{
mat_gen(dim)
}
\arguments{
\item{dim}{Integer specifying the dimension of the matrix.}
}
\value{
A symmetric matrix with dimensions \code{dim} x \code{dim}.
}
\description{
Creates a symmetric matrix with absolute random normal values.
}
\examples{
mat <- mat_gen(5)
}
