% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_weights.R
\name{missing_weights}
\alias{missing_weights}
\title{Calculate Missing Weights in a Network}
\usage{
missing_weights(W, validate = TRUE)
}
\arguments{
\item{W}{A matrix representing the network, where rows and columns correspond to
nodes, and the matrix elements represent the weight of the connections between
these nodes.}

\item{validate}{Whether to validate the input matrix.}
}
\value{
The calculated missing weights in the network.
}
\description{
Calculates the missing weights in a network by first validating the
input matrix, computing both inter-node and intra-node weights, summing these
weights to find the total (rounded to one significant figure), and finally
determining the missing weights by subtracting the sum of inter and intra-node
weights from the total.
}
\examples{
# Example matrix representing weighted connections between nodes
W <- matrix(c(0, 2, 1, 4, 2, 0, 3, 5, 1, 3, 0, 6, 4, 5, 6, 0), nrow = 4, byrow = TRUE)
missing_weights(W)
}
