% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_strength.R
\name{node_strength}
\alias{node_strength}
\title{Calculate the Strength of Nodes in an Undirected Network}
\usage{
node_strength(W)
}
\arguments{
\item{W}{A square, numeric matrix representing an undirected, weighted connection matrix of the network.
The elements of the matrix represent the weights of the edges between nodes. The matrix must be symmetric.}
}
\value{
A numeric vector containing the strength of each node.
}
\description{
This function computes the node strength in an undirected network,
defined as the sum of weights of edges connected to each node.
}
\examples{
# Create a 3x3 undirected, weighted connection matrix
W <- matrix(c(0, 2, 1, 4, 2, 0, 3, 5, 1, 3, 0, 6, 4, 5, 6, 0), nrow = 4, byrow = TRUE)
node_strength(W)
}
