% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalised_characteristic_path_length.R
\name{normalised_characteristic_path_length}
\alias{normalised_characteristic_path_length}
\title{Normalized Characteristic Path Length Calculation}
\usage{
normalised_characteristic_path_length(mat_list, rand_array = NULL)
}
\arguments{
\item{mat_list}{Either a square matrix representing the network for which the characteristic path length is to be calculated, or a list containing the network matrix and an optional 3D array of randomized matrices for comparison.}

\item{rand_array}{An optional 3D array of randomized matrices for comparison if the input is a single matrix. If NULL and the input is a matrix,
a set of randomized matrices is generated. This parameter is ignored if the input is a list.}
}
\value{
The normalized characteristic path length, computed as the ratio of the characteristic path length of the input network to the average characteristic path length of the randomized networks.
}
\description{
Computes the normalized characteristic path length for a given network, which can be provided either as a matrix or within a list alongside randomized matrices.
This function first checks if the input is a matrix or a list. If no array of randomized matrices is provided (or not included in the list), one is generated.
The function then calculates the average characteristic path length for the random array and normalizes the characteristic path length of the input network by this average.
The list elements must not be named.
}
\details{
Calculate Normalized Characteristic Path Length
}
\examples{
# Example for a single matrix input
W <- matrix(c(0, 2, 1, 0, 2, 0, 3, 5, 1, 3, 0, 6, 0, 5, 6, 0), nrow = 4, byrow = TRUE)
norm_char_path_length <- normalised_characteristic_path_length(W)
W_rand <- generateRewiredMatrices(W, 100)
W_rand <- abind::abind(W_rand, along = 3)
norm_char_path_length <- normalised_characteristic_path_length(list(W, W_rand))
# Can also be called with the following:
norm_char_path_length <- normalised_characteristic_path_length(W, W_rand)
}
