% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalised_clustering_coefficient.R
\name{normalised_clustering_coefficient}
\alias{normalised_clustering_coefficient}
\title{Normalized Clustering Coefficient Calculation}
\usage{
normalised_clustering_coefficient(mat_list, rand_array = NULL)
}
\arguments{
\item{mat_list}{A square matrix or a list containing a matrix and an optional 3D array of randomized matrices for comparison.
If provided as a list, the first element should be the matrix, and the second element (if present) should be the array of randomized matrices.
If a single matrix is provided and no array of randomized matrices is given, a set of randomized matrices is generated.}

\item{rand_array}{An optional 3D array of randomized matrices for comparison if the input is a single matrix. If NULL and the input is a matrix,
a set of randomized matrices is generated. This parameter is ignored if the input is a list.}
}
\value{
The normalized clustering coefficient, computed as the ratio of the clustering coefficient of the input matrix to the average clustering
coefficient of the randomized matrices. Returns 0 if the global clustering coefficient of the input is 0 to prevent division by zero errors.
}
\description{
Computes the normalized clustering coefficient for a given matrix.
The normalized clustering coefficient is calculated by dividing the global clustering coefficient of the input matrix (using \code{global_clustering_coefficient_wei()})
by the mean global clustering coefficient of a set of random matrices with a matched strength distribution. The random matrices can be generated by \code{generateRewiredMatrices()}.
Accepts either a numeric matrix or a list containing a matrix and array of random matrices already generated. If no array of randomized matrices is provided and the input is a matrix, the function generates one.
If the global clustering coefficient of the input is 0, the function returns 0 to avoid division by zero errors.
The list elements must not be named.
}
\details{
Calculate Normalized Clustering Coefficient for Single or List of Matrices
}
\examples{
# For a single matrix input
W <- matrix(c(0, 2, 1, 0, 2, 0, 3, 5, 1, 3, 0, 6, 0, 5, 6, 0), nrow = 4, byrow = TRUE)
norm_clust_coeff <- normalised_clustering_coefficient(W)
rand_matrices <- generateRewiredMatrices(W, 100)
rand_matrices <- abind::abind(rand_matrices, along = 3)
mat_list <- list(W, rand_matrices)
norm_clust_coeff <- normalised_clustering_coefficient(mat_list)
# Can also be called with the format:
norm_clust_coeff <- normalised_clustering_coefficient(W, rand_matrices)
}
