% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_matrices.R
\name{process_matrices}
\alias{process_matrices}
\title{Process and analyse matrices from files within a directory}
\usage{
process_matrices(
  directory,
  subjects_specified = NULL,
  file_convention,
  output_edge = FALSE
)
}
\arguments{
\item{directory}{The path to the directory containing the matrix files.}

\item{subjects_specified}{A vector of subject identifiers to be processed.}

\item{file_convention}{The naming convention for the matrix files related to each subject.}

\item{output_edge}{A boolean whether to calculate and output a dataframe containing the edge strengths of all connections in all networks.}
}
\value{
A list containing three elements: 'edge_df' which is a dataframe of edge strengths between lower
triangular matrix elements across all subjects, 'matrices' which is a 3D array of matrices,
where the matrices of different subjects are stacked in the Z axis, and 'subjects' which is a vector
containing the names of subjects who have matrices included in the dataset.
}
\description{
This function processes matrices stored in CSV files within a specified directory,
filtering by subject presence and adhering to a specified file naming convention. It validates
file existence, dimensions consistency, and generates an edge dataframe and an array of matrices.
}
\examples{
data_dir <- system.file("extdata", package = "NetAnalyseR")
subjects <- c("A", "B", "C", "D")
file_convention <- ".csv"
process_matrices(data_dir, subjects, file_convention)

}
