% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats2Table_fs7.R
\name{processFSdata}
\alias{processFSdata}
\title{Process MRI data from FreeSurfer output files}
\usage{
processFSdata(datasetDir, subject_names = NULL)
}
\arguments{
\item{datasetDir}{The directory containing the data files for all subjects.}

\item{subject_names}{A vector specifying the subject names to be analysed within the \code{datasetDir}.}
}
\value{
Returns a dataframe containing FreeSurfer brain volumes
}
\description{
This function reads, processes, and returns volumetric data from FreeSurfer output files for all subjects in a specified directory.
It reads volumetric and structural data from aseg.stats files, metadata from these files, area and thickness measurements from aparc.stats files,
and combines all these into a single dataset.
}
