% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_connectivity.R
\name{self_connectivity}
\alias{self_connectivity}
\title{Calculate Self-Connection Weights from a Network Matrix}
\usage{
self_connectivity(W, validate = TRUE)
}
\arguments{
\item{W}{A square, numeric matrix representing a weighted connection matrix of the network.}

\item{validate}{Whether to validate the input matrix.
Self-connections are represented by the diagonal elements of this matrix.}
}
\value{
A numeric vector containing the weights of self-connections for each node in the network.
}
\description{
This function returns the weights of self-connections within a network,
where each self-connection is a connection from a node to itself. It directly extracts
the diagonal of the input matrix, which represents these self-connections.
}
\examples{
# Create a 3x3 matrix with self-connections
W <- matrix(c(0, 2, 1, 4, 2, 0, 3, 5, 1, 3, 0, 6, 4, 5, 6, 0), nrow = 4, byrow = TRUE)
self_connectivity(W)
}
