% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_density.R
\name{threshold_density}
\alias{threshold_density}
\title{Threshold Density Function}
\usage{
threshold_density(mat, target_density, validate = TRUE)
}
\arguments{
\item{mat}{A numeric matrix representing the network.}

\item{target_density}{A numeric value indicating the target density.}

\item{validate}{A bool whether to validate the input matrix.}
}
\value{
A list containing the thresholded matrix (\code{matrix}), the applied threshold (\code{threshold}), and the resulting density (\code{density}).
}
\description{
Applies a threshold to a matrix to achieve a specified network density.
}
\examples{
mat <- mat_gen(100)
target_density <- 0.2
result <- threshold_density(mat, target_density)
image(result)

}
