% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{threshold_mat}
\alias{threshold_mat}
\title{Apply Threshold to Matrix}
\usage{
threshold_mat(mat, threshold)
}
\arguments{
\item{mat}{A numeric matrix.}

\item{threshold}{Numeric value to threshold matrix entries.}
}
\value{
A matrix with values below the threshold set to zero.
}
\description{
Applies a threshold to a matrix, setting elements below the threshold to zero.
}
\examples{
thresholded_mat <- threshold_mat(mat_gen(5), 50)
}
