% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{update_progress}
\alias{update_progress}
\title{Update Progress of Matrix Processing}
\usage{
update_progress(
  slice_counter,
  total_slices,
  start_time,
  metric_name,
  max_metric_length
)
}
\arguments{
\item{slice_counter}{Integer specifying the number of matrices processed so far.}

\item{total_slices}{Integer specifying the total number of matrices to process.}

\item{start_time}{POSIXct object indicating the start time of the processing.}

\item{metric_name}{Character string specifying the name of the metric being processed.}
}
\value{
None. Prints progress information to the console.
}
\description{
Displays the progress of processing matrices, including elapsed time, remaining time, and estimated time of completion.
}
\examples{
start_time <- Sys.time()
for (i in 1:100) {
  Sys.sleep(0.1)
  update_progress(i, 100, start_time, "Example Metric", 0)
}
}
