% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich_club.R
\name{compare_rich_club}
\alias{compare_rich_club}
\title{Rich Club Coefficient Comparison}
\usage{
compare_rich_club(
  mat_array,
  subjects,
  normalize = T,
  grouping_list,
  plot = FALSE,
  permutation = FALSE,
  n_rand = 100,
  p_adj = TRUE
)
}
\arguments{
\item{mat_array}{An array of connectivity matrices where each slice corresponds to a different subject.}

\item{subjects}{A vector of subject identifiers.}

\item{normalize}{Whether to compare rich club coefficient or normalised rich club coefficient between groups.}

\item{grouping_list}{A list specifying how subjects should be grouped for analysis.}

\item{plot}{A boolean indicating whether to generate plots of the results.}

\item{permutation}{A boolean to choose whether permutation tests should be used instead of ANOVA for statistical comparison.}

\item{n_rand}{The number of random permutations to use if permutation is TRUE.}

\item{p_adj}{A boolean indicating whether significant differences between groups on the graph should be controlled for multiple comparisons.}
}
\value{
A list containing statistical analysis results and optionally a plot, depending on the input parameters.
}
\description{
Computes and compares rich club coefficients for multiple subjects grouped by given criteria.
}
