% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_structural_deviation.R
\name{compute_network_deviation}
\alias{compute_network_deviation}
\title{Compute Network Deviations}
\usage{
compute_network_deviation(
  data_frame,
  matrices_array,
  control_group,
  threshold = 3
)
}
\arguments{
\item{data_frame}{A data frame containing at least two columns: \code{group} and \code{subject} for participant identifiers. The}

\item{matrices_array}{A 3D array where each slice corresponds to a matrix for a specific participant.}

\item{control_group}{A character specifying the level of \code{group} which should be considered the control group.}

\item{threshold}{A numeric value defining the Z-score threshold for considering deviations as significant.}
}
\value{
A data frame that joins the input data_frame with computed network deviations for each case.
}
\description{
This function calculates network deviations for subjects based on the provided control and subject groups within a data frame.
It extracts control and subject matrices from a 3D array based on group classification in the data frame and then calculates network deviations
using a specified threshold.
}
\details{
Compute network deviations for a given dataset and matrices array
}
\examples{
data_dir <- system.file("extdata", package = "NetAnalyseR")
subjects <- c("A", "B", "C", "D")
file_convention <- ".csv"
output <- process_matrices(data_dir, subjects, file_convention)
global_metrics <- c("characteristic_path_length", "global_efficiency_wei")
global_df <- compute_global_metrics(matrices_array = output$matrices,
                                    global_metrics = global_metrics,
                                    subject_names = output$subjects)
grouping_list <- list(
  control = c("A", "B"),
  treatment = c("C", "D")
)
allocated_df <- allocate_groups(global_df, grouping_list)
global_deviation <- compute_network_deviation(allocated_df, output$matrices, "control")
}
\references{
Gugger, J. J., Sinha, N., Huang, Y., Walter, A. E., Lynch, C., Kalyani, P., Smyk, N., Sandsmark, D., Diaz-Arrastia, R., & Davis, K. A. (2023). Structural brain network deviations predict recovery after traumatic brain injury. NeuroImage clinical, 38, 103392-103392. https://doi.org/10.1016/j.nicl.2023.103392
}
