% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateRewiredMatrices}
\alias{generateRewiredMatrices}
\title{Generate a Series of Rewired Matrices}
\usage{
generateRewiredMatrices(initialMatrix, n = 100L)
}
\arguments{
\item{initialMatrix}{A numeric matrix representing the initial network's adjacency matrix.}

\item{n}{The number of rewired matrices to generate.}
}
\value{
A list of 'n' rewired matrices.
}
\description{
Rewires a matrix n times, performing 10 iterations each time.
Saves all random matrices generated as a list.
The output is in the format of a list. If manually using the output in normalised mesures
the output must be converted to an array with \code{abind::abind(matrix, along = 3)}
}
