% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_comparisons.R
\name{group_comparisons}
\alias{group_comparisons}
\title{Group Comparisons for Metrics with Dynamic Testing and Visualization}
\usage{
group_comparisons(
  group_df,
  metrics,
  comparisons,
  p_adjust_method = "BH",
  group_order = NULL,
  test_type = NULL,
  num_permutations = 1000,
  analysis_type = "between",
  id_var = NULL,
  time_var = NULL
)
}
\arguments{
\item{group_df}{(Dataframe) containing 'subject', 'group', and metric columns.}

\item{metrics}{(Character) vector of metric column names to analyze.}

\item{comparisons}{(List) of pairwise group comparisons for tests or 'all' to include all possible pairwise comparisons.}

\item{p_adjust_method}{(Character) string specifying the p-value adjustment method (default is "BH").}

\item{group_order}{Optional (Character) vector specifying the order of groups.}

\item{test_type}{(Character) string specifying "parametric" or "nonparametric" (optional). If NULL, test type is determined by normality and homogeneity tests.}

\item{num_permutations}{(Integer) specifying how many iterations should be used for the permuation test.}

\item{id_var}{(Character) The variable which groups observations into a single entity.}

\item{time_var}{(Character) The variable which delineates observations within a single entity.}
}
\value{
(List) with overall test results, pairwise test results, normality and homogeneity results, and a combined plot.
}
\description{
Performs statistical tests on specified metrics across groups, adjusting p-values and generating combined plots. Supports dynamic determination of test type.
}
