% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_inversion.R
\name{length_inversion}
\alias{length_inversion}
\title{Convert Matrix Weights to Lengths}
\usage{
length_inversion(W)
}
\arguments{
\item{W}{A numeric matrix whose weights represent connection strengths.}
}
\value{
A matrix where each non-zero weight has been converted to its reciprocal value, representing the length.
}
\description{
Converts the weights in a numeric matrix to lengths based on the principle that length is
inversely proportional to the connection strength. This function first validates the input matrix,
sets its diagonal elements to zero to avoid self-connections, and then converts all non-zero
weights to their reciprocal values, effectively transforming weights into lengths.
}
\examples{
W <- matrix(c(0, 2, 1, 4, 2, 0, 3, 5, 1, 3, 0, 6, 4, 5, 6, 0), nrow = 4, byrow = TRUE)
lengths_W <- length_inversion(W)

}
