% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rewireNetworkCpp}
\alias{rewireNetworkCpp}
\title{Rewire a Network Matrix}
\usage{
rewireNetworkCpp(R, initialIter = 2L, validate = TRUE)
}
\arguments{
\item{R}{A numeric matrix representing the network's adjacency matrix.}

\item{initialIter}{int number of iterations for the rewiring process.}

\item{validate}{bool specifying whether to check the validity of the input matrix}
}
\value{
A rewired numeric matrix representing the network's adjacency matrix.
}
\description{
This function takes a network represented as an adjacency matrix and
performs a specified number of edge rewiring iterations to randomize the network.
It checks for network density and skips rewiring if the network is fully connected.
}
