% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_connectivity.R
\name{self_connectivity}
\alias{self_connectivity}
\title{Calculate Self-Connection Weights from a Network Matrix}
\usage{
self_connectivity(W)
}
\arguments{
\item{W}{A square, numeric matrix representing a weighted connection matrix of the network.
Self-connections are represented by the diagonal elements of this matrix.}
}
\value{
A numeric vector containing the weights of self-connections for each node in the network.
}
\description{
This function returns the weights of self-connections within a network,
where each self-connection is a connection from a node to itself. It directly extracts
the diagonal of the input matrix, which represents these self-connections.
}
\examples{
# Create a 3x3 matrix with self-connections
W <- matrix(c(1, 2, 3,
              4, 5, 6,
              7, 8, 9), byrow = TRUE, nrow = 3)
self_connectivity(W)

}
