% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_matrix.R
\name{validate_matrix}
\alias{validate_matrix}
\title{Validate a Square Numeric Matrix}
\usage{
validate_matrix(W, skip = FALSE)
}
\arguments{
\item{W}{A matrix to validate.}
}
\value{
Returns a valid symmetric square numeric matrix.
If the matrix fails any validation checks, the function stops with an appropriate error message.
}
\description{
This function checks if the input is a valid square numeric matrix without NA values,
ensuring it is symmetric. If the matrix is not symmetric, it modifies the matrix
to make it symmetric by mirroring the lower triangular part onto the upper triangular part.
}
\keyword{internal}
